Refactor: chnage directory Lerning
