+++ src/components/Learning/components/VideoPlayer/VideoPlayer.component.tsx
