+++ src/components/Learning/components/VideoSidebar/ChapterAccordion/ChapterAccordion.component.tsx
